Added DelegateInstallConvention and applied it to TransferPackageCommand
